Updated README and data files
type: chore

Documented the purpose of the project and referenced the appropriate
packages. This is a fairly manual bit of code which can be simplified
through automatic downloads of the spatial files, but explained the
current process. Also forgot to include the data folder and code
to make it run.