Add Troubleshoot for rJava
docs: Add troubleshoot when rJava enviroment is not found

- Document clearly the root cause for rJava not running correctly is
- Guide users through steps to install Java based on the version of
R being run

There is a common issue with R and Java eniroments installed with
different bit systems, which are incompatible with one another. The
script relies on an Excel editor package (xlsx) whose dependancy is
rJava. When a user is working in a 32-bit version of R, the only way
rJava can run successfully in xlsx is when there is a 32-bit version of
Java installed. Same goes for the 64-bit version of R, which requires
a 64-bit version of Java.