chore: Add Sample Data for Script
- Provide sample data for new users to get familiar with the code
- Build demo code around sample data to see how to work with it

Starting in geospatial analysis is tricky because there are a lot of new
kinds of data one needs to understand just to start. Working with shape
files or wrangling together custom zones takes a lot of knowledge about
said shape file, so some sample data and code can really help a user
understand the 'how' of this code. It also simplifies the state and
county identification process and how to actually color-code the heat
map.